29951 ProtocolData vocabulary not exposed Added test for vocabulary protocolData, moved vocabulary based configuraiton class to config package.
