+++ src/main/java/ca/gc/aafc/collection/api/config/CollectionExtensionConfiguration.java
@@ -1,4 +1,4 @@
-package ca.gc.aafc.collection.api;
+package ca.gc.aafc.collection.api.config;
 
 import java.util.Map;
