+++ src/main/java/ca/gc/aafc/collection/api/validation/CollectingEventExtensionValueValidator.java
@@ -3,7 +3,7 @@ package ca.gc.aafc.collection.api.validation;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Component;
 
-import ca.gc.aafc.collection.api.CollectionExtensionConfiguration;
+import ca.gc.aafc.collection.api.config.CollectionExtensionConfiguration;
 import ca.gc.aafc.collection.api.entities.DinaComponent;
 
 @Component
