+++ src/test/java/ca/gc/aafc/collection/api/ExtensionConfigurationTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import javax.inject.Inject;
 
+import ca.gc.aafc.collection.api.config.CollectionExtensionConfiguration;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
